Refactoring and Additional Endpoints for Customers and Products

+++ src/main/java/group/csci318groupproj/repository/CustomerRepository.java
@@ -3,6 +3,8 @@ package group.csci318groupproj.repository;
 import group.csci318groupproj.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
+//Interface that extends the JPA library for H2 functionality for Customer Class
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 
 }
