+++ src/main/java/group/csci318groupproj/repository/ProductRepository.java
@@ -3,6 +3,8 @@ package group.csci318groupproj.repository;
 import group.csci318groupproj.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+
+//Interface that extends the JPA library for H2 functionality for Product Class
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
 }
